# app/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
paths:
...
  /jobs:
    get:
      summary: Get paginated list of jobs
      description: Fetch a paginated list of jobs sorted by ID in ascending order.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination.
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
          description: Number of items per page.
      responses:
        200:
          description: Successful response with jobs data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        company_name:
                          type: string
                        salary:
                          type: string
                        deadline:
                          type: string
                          format: date
                  pagination:
                    type: object
                    properties:
                      current_page:
                        type: integer
                      per_page:
                        type: integer
        500:
          description: Server error

    post:
      summary: Create a new job
      description: Create a new job with the given details.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Software Engineer
                company:
                  type: string
                  example: Example Corp
                company_link:
                  type: string
                  example: https://example.com
                location:
                  type: string
                  example: example location
                salary:
                  type: string
                  example: 6000
                career:
                  type: string
                  example: example career
                education:
                  type: string
                  example: example education
                employment:
                  type: string
                  example: example employment
                deadline:
                  type: string
                  format: date
                  example: 2024-12-31
                tags:
                  type: array
                  items:
                    type: string
                  example: ["example tag"]
                link:
                  type: string
                  example: https://jobportal.com/job123
      responses:
        201:
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Job created successfully
                  job_id:
                    type: integer
                    example: 1
        400:
          description: Validation error
        500:
          description: Server error
  /jobs/{job_id}:
    get:
      summary: Get job details
      description: Retrieve detailed information about a specific job, including its company, location, and tags.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - name: job_id
          in: path
          description: The ID of the job to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successfully retrieved job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: Backend Developer
                      salary:
                        type: string
                        example: 1000000
                      career:
                        type: string
                        example: 3 years
                      education:
                        type: string
                        example: Bachelor's
                      employment:
                        type: string
                        example: Full-time
                      deadline:
                        type: string
                        format: date
                        example: 2024-12-31
                      link:
                        type: string
                        example: http://job.example.com/1
                      company:
                        type: object
                        properties:
                          name:
                            type: string
                            example: Tech Corp
                          link:
                            type: string
                            example: http://techcorp.example.com
                      location:
                        type: object
                        properties:
                          region:
                            type: string
                            example: Seoul
                          district:
                            type: string
                            example: Gangnam-gu
                      tags:
                        type: array
                        items:
                          type: string
                        example:
                          - Backend
                          - Python
                          - Django
        404:
          description: Job not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Job not found
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
    put:
      summary: Update an existing job
      description: Update details of an existing job posting by ID.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: ID of the job to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: example new title
                company:
                  type: string
                  example: example new company name
                company_link:
                  type: string
                  example: http://newcompany.example.com
                location:
                  type: string
                  example: example new location
                salary:
                  type: string
                  example: example new salary
                career:
                  type: string
                  example: example new career
                education:
                  type: string
                  example: example new education
                employment:
                  type: string
                  example: example new employment
                deadline:
                  type: string
                  format: date
                  example: 2024-12-31
                tags:
                  type: array
                  items:
                    type: string
                    example: ['example new tag']
                link:
                  type: string
                  example: http://newjob.example.com/1
      responses:
        200:
          description: Job updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Job updated successfully
        400:
          description: Validation error
        404:
          description: Job not found
        500:
          description: Server error

    delete:
      summary: Delete a job
      description: Deletes a job posting by ID and all its related data.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: ID of the job to delete
      responses:
        200:
          description: Job deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Job deleted successfully
        404:
          description: Job not found
        500:
          description: Server error
  /jobs/search:
    get:
      summary: Search jobs by keyword
      description: Search for jobs using a keyword. The keyword will be matched against job title, company name, and description.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
          description: The keyword to search for
        - in: query
          name: page
          schema:
            default: 1
            type: integer
          description: Page number for pagination
        - in: query
          name: per_page
          schema:
            default: 20
            type: integer
          description: Number of results per page
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        company_name:
                          type: string
                        salary:
                          type: string
                        deadline:
                          type: string
                          format: date
        400:
          description: Validation error
        500:
          description: Server error
  /jobs/filter:
    get:
      summary: Filter jobs by location and tags
      description: Filter jobs by location and tags. At least one location or tag must be provided.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - in: query
          name: location
          schema:
            type: array
            items:
              type: string
            description: List of locations to filter by
        - in: query
          name: tag
          schema:
            type: array
            items:
              type: string
            description: List of tags to filter by
        - in: query
          name: page
          schema:
            default: 1
            type: integer
          description: Page number for pagination
        - in: query
          name: per_page
          schema:
            default: 20
            type: integer
          description: Number of results per page
      responses:
        200:
          description: Filter results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        company_name:
                          type: string
                        salary:
                          type: string
                        deadline:
                          type: string
                          format: date
        400:
          description: Validation error
        500:
          description: Server error
  /jobs/sort:
    get:
      summary: Sort jobs by deadline
      description: Sort jobs by their deadline in ascending or descending order.
      tags:
        - Jobs
      security:
        - BearerAuth: [ ]
      parameters:
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            description: Order to sort the jobs by (ascending or descending)
        - in: query
          name: page
          schema:
            default: 1
            type: integer
          description: Page number for pagination
        - in: query
          name: per_page
          schema:
            default: 20
            type: integer
          description: Number of results per page
      responses:
        200:
          description: Sorted results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        company_name:
                          type: string
                        salary:
                          type: string
                        deadline:
                          type: string
                          format: date
        400:
          description: Validation error
        500:
          description: Server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT